refactor: extract pure utils and factories into standalone modules
Extract dependency-free modules (questionFactory, groupFactory, pureUtils,
surveyTheme) from existing files to improve code organization. Original
files re-export for backward compatibility. No functional changes.
