+++ src/components/design/NewComponentsPanel/groupFactory.js
@@ -1,0 +1,13 @@
+export const createGroup = (groupType, gId) => {
+  let code = `G${gId}`;
+  let state = {
+    groupType,
+  };
+  let newGroup = {
+    code,
+    qualifiedCode: code,
+    type: groupType.toLowerCase(),
+    groupType,
+  };
+  return { newGroup, state };
+};
