+++ src/components/design/NewComponentsPanel/index.jsx
@@ -93,16 +93,4 @@ function NewComponentsPanel({ t }) {
 
 export default React.memo(NewComponentsPanel);
 
-export const createGroup = (groupType, gId) => {
-  let code = `G${gId}`;
-  let state = {
-    groupType,
-  };
-  let newGroup = {
-    code,
-    qualifiedCode: code,
-    type: groupType.toLowerCase(),
-    groupType,
-  };
-  return { newGroup, state };
-};
+export { createGroup } from "./groupFactory";
