+++ src/constants/surveyTheme.js
@@ -1,0 +1,33 @@
+export const BG_COLOR = "#dfe2ef";
+export const TEXT_COLOR = "#091133";
+export const PRIMARY_COLOR = "#2d3cb1";
+export const SECONDARY_COLOR = "#672ebf";
+export const ERR_COLOR = "#d13e17";
+export const PAPER_COLOR = "#ffffff";
+export const GROUP_FONT_SIZE = 32;
+export const QUESTION_FONT_SIZE = 18;
+export const TEXT_FONT_SIZE = 14;
+export const FONT_FAMILY = "Rubik";
+
+export const defaultSurveyTheme = {
+  textStyles: {
+    group: {
+      font: FONT_FAMILY,
+      size: GROUP_FONT_SIZE,
+      color: TEXT_COLOR,
+    },
+    question: {
+      font: FONT_FAMILY,
+      size: QUESTION_FONT_SIZE,
+      color: TEXT_COLOR,
+    },
+    text: {
+      font: FONT_FAMILY,
+      size: TEXT_FONT_SIZE,
+      color: TEXT_COLOR,
+    },
+  },
+  primaryColor: PRIMARY_COLOR,
+  bgColor: BG_COLOR,
+  paperColor: PAPER_COLOR,
+};
