+++ src/constants/theme.js
@@ -1,15 +1,13 @@
 import { colorToThemeMode } from "~/components/Questions/utils";
-
-export const BG_COLOR = "#dfe2ef";
-export const TEXT_COLOR = "#091133";
-export const PRIMARY_COLOR = "#2d3cb1";
-export const SECONDARY_COLOR = "#672ebf";
-export const ERR_COLOR = "#d13e17";
-export const PAPER_COLOR = "#ffffff";
-export const GROUP_FONT_SIZE = 32;
-export const QUESTION_FONT_SIZE = 18;
-export const TEXT_FONT_SIZE = 14;
-export const FONT_FAMILY = "Rubik";
+export {
+  BG_COLOR, TEXT_COLOR, PRIMARY_COLOR, SECONDARY_COLOR, ERR_COLOR,
+  PAPER_COLOR, GROUP_FONT_SIZE, QUESTION_FONT_SIZE, TEXT_FONT_SIZE, FONT_FAMILY,
+  defaultSurveyTheme,
+} from "./surveyTheme";
+import {
+  BG_COLOR, TEXT_COLOR, PRIMARY_COLOR, SECONDARY_COLOR, ERR_COLOR,
+  PAPER_COLOR, GROUP_FONT_SIZE, QUESTION_FONT_SIZE, TEXT_FONT_SIZE, FONT_FAMILY,
+} from "./surveyTheme";
 
 export const defualtTheme = (theme) => {
   return {
@@ -55,27 +53,4 @@ export const defualtTheme = (theme) => {
       },
     },
   };
-};
-
-export const defaultSurveyTheme = {
-  textStyles: {
-    group: {
-      font: FONT_FAMILY,
-      size: GROUP_FONT_SIZE,
-      color: TEXT_COLOR,
-    },
-    question: {
-      font: FONT_FAMILY,
-      size: QUESTION_FONT_SIZE,
-      color: TEXT_COLOR,
-    },
-    text: {
-      font: FONT_FAMILY,
-      size: TEXT_FONT_SIZE,
-      color: TEXT_COLOR,
-    },
-  },
-  primaryColor: PRIMARY_COLOR,
-  bgColor: BG_COLOR,
-  paperColor: PAPER_COLOR,
 };
