+++ src/utils/design/utils.js
@@ -6,43 +6,13 @@ import {
   STRIP_TAGS_PATTERN,
 } from "~/constants/instruction";
 
-export const isEquivalent = (a, b, visited = new WeakSet()) => {
-  if (a === b) return true;
-
-  if (typeof a === "function" || typeof b === "function") {
-    return false;
-  }
-
-  if (typeof a !== "object" || typeof b !== "object") {
-    return a === b;
-  }
-
-  if (a === null || b === null) {
-    return a === b;
-  }
-
-  if (visited.has(a) || visited.has(b)) {
-    return true;
-  }
-
-  visited.add(a);
-  visited.add(b);
-
-  const aProps = Object.getOwnPropertyNames(a);
-  const bProps = Object.getOwnPropertyNames(b);
-
-  if (aProps.length !== bProps.length) {
-    return false;
-  }
-
-  for (const prop of aProps) {
-    if (prop !== "key" && !isEquivalent(a[prop], b[prop], visited)) {
-      return false;
-    }
-  }
-
-  return true;
-};
+export {
+  isEquivalent,
+  nextId,
+  firstIndexInArray,
+  lastIndexInArray,
+} from "./pureUtils";
+import { isEquivalent } from "./pureUtils";
 
 export const diff = (obj1, obj2) => {
   if (!obj2 || Object.prototype.toString.call(obj2) !== "[object Object]") {
@@ -123,26 +93,6 @@ export const diff = (obj1, obj2) => {
   return diffs;
 };
 
-export const nextId = (elements) => {
-  if (elements.length) {
-    let arrayOfIntCodes = elements
-      .filter((el) => el.type != "other")
-      .map((el) => el.code.replace(/^\D+/g, ""))
-      .filter((el) => el.length > 0);
-    if (arrayOfIntCodes.length) {
-      let intCodes = arrayOfIntCodes
-        .map((el) => parseInt(el, 10))
-        .sort(function (a, b) {
-          return a - b;
-        });
-      if (intCodes) {
-        return intCodes[intCodes.length - 1] + 1;
-      }
-    }
-  }
-  return 1;
-};
-
 export const stripTags = (string) => {
   if (typeof string !== "string") return string;
   return string
@@ -160,31 +110,6 @@ export function truncateWithEllipsis(text, maxLength) {
 
 export const isQuestion = (code) => QUESTION_CODE_PATTERN.test(code);
 export const isGroup = (code) => GROUP_CODE_PATTERN.test(code);
-
-export const lastIndexInArray = (array, func) => {
-  if (!array) {
-    return -1;
-  }
-  let index = array.length - 1;
-  for (; index >= 0; index--) {
-    if (func(array[index])) {
-      return index;
-    }
-  }
-  return -1;
-};
-
-export const firstIndexInArray = (array, func) => {
-  if (!array) {
-    return -1;
-  }
-  for (let index = 0; index < array.length; index++) {
-    if (func(array[index])) {
-      return index;
-    }
-  }
-  return -1;
-};
 
 export const isNotEmptyHtml = (value) => {
   if (!value) return false;
